requirements.txt
Flask
boto3
flask-cors
python-dotenv